+++ contracts/mocks/BaseMockLiquidation.sol
@@ -2,10 +2,10 @@
 pragma solidity >0.7.0;
 pragma experimental ABIEncoderV2;
 
-import "../storage/PortfolioHandler.sol";
-import "../storage/AccountContextHandler.sol";
-import "../common/Liquidation.sol";
-import "../storage/StorageLayoutV1.sol";
+import "../internal/portfolio/PortfolioHandler.sol";
+import "../internal/AccountContextHandler.sol";
+import "../internal/Liquidation.sol";
+import "../global/StorageLayoutV1.sol";
 
 contract BaseMockLiquidation is StorageLayoutV1 {
     using PortfolioHandler for PortfolioState;
@@ -13,32 +13,25 @@ contract BaseMockLiquidation is StorageLayoutV1 {
     using Liquidation for LiquidationFactors;
     using Market for MarketParameters;
 
-    function setAssetRateMapping(
-        uint id,
-        AssetRateStorage calldata rs
-    ) external {
+    function setAssetRateMapping(uint256 id, AssetRateStorage calldata rs) external {
         assetToUnderlyingRateMapping[id] = rs;
     }
 
-    function setCashGroup(
-        uint id,
-        CashGroupParameterStorage calldata cg
-    ) external {
+    function setCashGroup(uint256 id, CashGroupParameterStorage calldata cg) external {
         CashGroup.setCashGroupStorage(id, cg);
     }
 
-    function buildCashGroupView(
-        uint currencyId
-    ) public view returns (
-        CashGroupParameters memory,
-        MarketParameters[] memory
-    ) {
+    function buildCashGroupView(uint256 currencyId)
+        public
+        view
+        returns (CashGroupParameters memory, MarketParameters[] memory)
+    {
         return CashGroup.buildCashGroupView(currencyId);
     }
 
     function setMarketStorage(
-        uint currencyId,
-        uint settlementDate,
+        uint256 currencyId,
+        uint256 settlementDate,
         MarketParameters memory market
     ) public {
         market.storageSlot = Market.getSlot(currencyId, settlementDate, market.maturity);
@@ -47,19 +40,14 @@ contract BaseMockLiquidation is StorageLayoutV1 {
         market.setMarketStorage();
     }
 
-    function setETHRateMapping(
-        uint id,
-        ETHRateStorage calldata rs
-    ) external {
+    function setETHRateMapping(uint256 id, ETHRateStorage calldata rs) external {
         underlyingToETHRateMapping[id] = rs;
     }
 
-    function setPortfolio(
-        address account,
-        PortfolioAsset[] memory assets
-    ) external {
+    function setPortfolio(address account, PortfolioAsset[] memory assets) external {
         AccountStorage memory accountContext = AccountContextHandler.getAccountContext(account);
-        PortfolioState memory portfolioState = PortfolioHandler.buildPortfolioState(account, accountContext.assetArrayLength, 0);
+        PortfolioState memory portfolioState =
+            PortfolioHandler.buildPortfolioState(account, accountContext.assetArrayLength, 0);
         portfolioState.newAssets = assets;
         accountContext.storeAssetsAndUpdateContext(account, portfolioState);
         accountContext.setAccountContext(account);
@@ -67,23 +55,27 @@ contract BaseMockLiquidation is StorageLayoutV1 {
 
     function setBalance(
         address account,
-        uint currencyId,
-        int cashBalance,
-        int perpTokenBalance
+        uint256 currencyId,
+        int256 cashBalance,
+        int256 perpTokenBalance
     ) external {
         AccountStorage memory accountContext = AccountContextHandler.getAccountContext(account);
-        accountContext.setActiveCurrency(currencyId, true, AccountContextHandler.ACTIVE_IN_BALANCES_FLAG);
+        accountContext.setActiveCurrency(
+            currencyId,
+            true,
+            AccountContextHandler.ACTIVE_IN_BALANCES_FLAG
+        );
         accountContext.setAccountContext(account);
 
         bytes32 slot = keccak256(abi.encode(currencyId, account, "account.balances"));
 
-        bytes32 data = (
-            (bytes32(uint(perpTokenBalance))) |
-            (bytes32(0) << 96) |
-            (bytes32(cashBalance) << 128)
-        );
+        bytes32 data =
+            ((bytes32(uint256(perpTokenBalance))) |
+                (bytes32(0) << 96) |
+                (bytes32(cashBalance) << 128));
 
-        assembly { sstore(slot, data) }
+        assembly {
+            sstore(slot, data)
+        }
     }
-
 }
