+++ contracts/mocks/MockAccountContextHandler.sol
@@ -2,57 +2,48 @@
 pragma solidity >0.7.0;
 pragma experimental ABIEncoderV2;
 
-import "../storage/AccountContextHandler.sol";
+import "../internal/AccountContextHandler.sol";
 
 contract MockAccountContextHandler {
     using AccountContextHandler for AccountStorage;
 
     function setAssetBitmap(
         address account,
-        uint id,
+        uint256 id,
         bytes32 bitmap
     ) external {
         BitmapAssetsHandler.setAssetsBitmap(account, id, bitmap);
     }
 
-    function enableBitmapForAccount(
-        address account,
-        uint currencyId
-    ) external {
+    function enableBitmapForAccount(address account, uint256 currencyId) external {
         AccountStorage memory accountContext = AccountContextHandler.getAccountContext(account);
         accountContext.enableBitmapForAccount(account, currencyId);
         accountContext.setAccountContext(account);
     }
 
-    function getAccountContext(
-        address account
-    ) external view returns (AccountStorage memory) {
+    function getAccountContext(address account) external view returns (AccountStorage memory) {
         return AccountContextHandler.getAccountContext(account);
     }
 
-    function setAccountContext(
-        AccountStorage memory accountContext,
-        address account
-    ) external {
+    function setAccountContext(AccountStorage memory accountContext, address account) external {
         return accountContext.setAccountContext(account);
     }
 
-    function isActiveInBalances(
-        AccountStorage memory accountContext,
-        uint currencyId
-    ) external pure returns (bool) {
+    function isActiveInBalances(AccountStorage memory accountContext, uint256 currencyId)
+        external
+        pure
+        returns (bool)
+    {
         return accountContext.isActiveInBalances(currencyId);
     }
 
-    function clearPortfolioActiveFlags(
-        bytes18 activeCurrencies
-    ) external pure returns (bytes18) {
+    function clearPortfolioActiveFlags(bytes18 activeCurrencies) external pure returns (bytes18) {
         return AccountContextHandler.clearPortfolioActiveFlags(activeCurrencies);
     }
 
     function setActiveCurrency(
         bytes18 activeCurrencies,
-        uint currencyId,
+        uint256 currencyId,
         bool isActive,
         bytes2 flags
     ) external pure returns (bytes18) {
@@ -61,22 +52,25 @@ contract MockAccountContextHandler {
 
         // Assert that the currencies are in order
         bytes18 currencies = accountContext.activeCurrencies;
-        uint lastCurrency;
+        uint256 lastCurrency;
         while (currencies != 0x0) {
-            uint thisCurrency = uint(uint16(bytes2(currencies) & AccountContextHandler.UNMASK_FLAGS));
-            assert (thisCurrency != 0);
+            uint256 thisCurrency =
+                uint256(uint16(bytes2(currencies) & AccountContextHandler.UNMASK_FLAGS));
+            assert(thisCurrency != 0);
             // Either flag must be set
-            assert (
-                ((bytes2(currencies) & AccountContextHandler.ACTIVE_IN_PORTFOLIO_FLAG) == AccountContextHandler.ACTIVE_IN_PORTFOLIO_FLAG)
-                || ((bytes2(currencies) & AccountContextHandler.ACTIVE_IN_BALANCES_FLAG) == AccountContextHandler.ACTIVE_IN_BALANCES_FLAG)
+            assert(
+                ((bytes2(currencies) & AccountContextHandler.ACTIVE_IN_PORTFOLIO_FLAG) ==
+                    AccountContextHandler.ACTIVE_IN_PORTFOLIO_FLAG) ||
+                    ((bytes2(currencies) & AccountContextHandler.ACTIVE_IN_BALANCES_FLAG) ==
+                        AccountContextHandler.ACTIVE_IN_BALANCES_FLAG)
             );
             // currencies are in order
-            assert (thisCurrency > lastCurrency);
+            assert(thisCurrency > lastCurrency);
 
             if (isActive && currencyId == thisCurrency) {
-                assert (bytes2(currencies) & flags == flags);
+                assert(bytes2(currencies) & flags == flags);
             } else if (!isActive && currencyId == thisCurrency) {
-                assert (bytes2(currencies) & flags != flags);
+                assert(bytes2(currencies) & flags != flags);
             }
 
             lastCurrency = thisCurrency;
@@ -85,5 +79,4 @@ contract MockAccountContextHandler {
 
         return accountContext.activeCurrencies;
     }
-
 }
